condevofm/atoms/calculate.py
"""
calculate.py

Implementation of calculation functions for ase.atoms.Atoms objects.
"""

from functools import partial
from multiprocessing import Pool

import numpy as np
import torch
from ase import Atoms
from ase.calculators.lj import LennardJones
from mace.calculators import MACECalculator
from tqdm import tqdm


def init_calc(calc_str, device="cpu"):
    calc = None
    if calc_str == "LJ":
        calc = LennardJones(sigma=1 / (2 ** (1.0 / 6.0)), rc=10.0, smooth=True)
    else:
        calc = MACECalculator(
            model_paths=calc_str,
            device=device,
            default_dtype="float64",
        )

    return calc


def sample_to_atoms(sample, free_atoms_template, free_indices=None):
    """
    Convert a 1D sample vector to an Atoms object of the free subset.

    free_indices is optional: if not provided, it will be read from
    free_atoms_template.info['indices'] when available.
    """

    # infer free_indices if not provided
    if free_indices is None:
        if (
            free_atoms_template.info is None
            or "indices" not in free_atoms_template.info
        ):
            raise ValueError(
                "free_indices must be provided either as an argument or present in free_atoms_template.info['indices']"
            )
        free_indices = free_atoms_template.info["indices"]

    n_free = len(free_atoms_template)
    reshaped = np.asarray(sample, dtype=float).reshape(n_free, 3)
    out = free_atoms_template.copy()
    out.set_positions(reshaped)
    # store indices order
    info = dict(out.info) if out.info is not None else {}
    info["indices"] = list(free_indices)
    out.info = info
    return out


def combine_fixed_frozen_and_free_atoms(
    founder_atoms,
    fixed_atoms,
    fixed_indices,
    free_atoms,
    frozen_atoms=None,
    frozen_indices=None,
    calc=None,
):
    """
    Reconstruct a full 'Atoms' object from free-subset 'free_atoms' by
    inserting frozen and fixed atoms at the supplied indices.

    Args:
        founder_atoms: full Atoms used for cell, pbc and as fallback.
        fixed_atoms: Atoms containing the fixed atoms (order must match
            'fixed_indices'). If None, fixed positions are taken from founder.
        fixed_indices: indices in the full Atoms.
        free_atoms: Atoms containing only the free subset positions (in
            the ordering that will match insertion into the full array).
        frozen_atoms: (optional) Atoms containing frozen atoms (order must
            match 'frozen_indices').
        frozen_indices: (optional) indices for frozen atoms in the full Atoms.
        calc: optional ASE calculator to attach to the returned Atoms.

    Returns:
        combined_atoms: full Atoms with positions reconstructed.
    """

    if free_atoms is None:
        raise ValueError("free_atoms must be provided")

    if frozen_indices is None:
        frozen_indices = []
    if fixed_indices is None:
        fixed_indices = []

    if frozen_atoms is not None and len(frozen_indices) != len(frozen_atoms):
        raise ValueError(
            "frozen_indices length must equal frozen_atoms length."
        )

    if fixed_atoms is not None and len(fixed_indices) != len(fixed_atoms):
        raise ValueError("fixed_indices length must equal fixed_atoms length.")

    n_atoms = len(founder_atoms)
    combined_positions = np.zeros((n_atoms, 3), dtype=float)

    def _assign_positions(indices, source_atoms):
        if not indices:
            return
        src_pos = np.asarray(source_atoms.get_positions())
        if src_pos.shape[0] != len(indices):
            raise ValueError(
                f"Source positions length ({src_pos.shape[0]}) does not match indices length ({len(indices)})"
            )
        for i, idx in enumerate(indices):
            combined_positions[int(idx)] = src_pos[i]

    if fixed_indices:
        if fixed_atoms is not None:
            _assign_positions(fixed_indices, fixed_atoms)
        else:
            founder_pos = founder_atoms.get_positions()
            for idx in fixed_indices:
                combined_positions[int(idx)] = founder_pos[int(idx)]

    if frozen_indices:
        if frozen_atoms is None:
            # if frozen_indices exist but no frozen_atoms object provided, take from founder
            founder_pos = founder_atoms.get_positions()
            for idx in frozen_indices:
                combined_positions[int(idx)] = founder_pos[int(idx)]
        else:
            _assign_positions(frozen_indices, frozen_atoms)

    # Assign free atoms: infer free_indices order if needed
    if free_atoms is not None:
        if (
            hasattr(free_atoms, "info")
            and free_atoms.info is not None
            and "indices" in free_atoms.info
        ):
            free_indices = free_atoms.info["indices"]
        elif "free_indices" in locals() and free_indices is not None:
            free_indices = free_indices
        else:
            raise ValueError(
                "free_indices must be provided either as an argument or present in free_atoms.info['indices']"
            )

        free_pos = np.asarray(free_atoms.get_positions())
        if free_pos.shape[0] != len(free_indices):
            raise ValueError(
                f"free_atoms positions length ({free_pos.shape[0]}) does not match free_indices length ({len(free_indices)})"
            )
        for i, idx in enumerate(free_indices):
            combined_positions[int(idx)] = free_pos[i]

    founder_pos = founder_atoms.get_positions()
    for i in range(n_atoms):
        if np.allclose(combined_positions[i], 0.0):
            combined_positions[i] = founder_pos[i]

    combined_atoms = Atoms(
        symbols=founder_atoms.get_chemical_symbols(),
        positions=combined_positions,
        cell=founder_atoms.get_cell(),
        pbc=founder_atoms.get_pbc(),
    )

    if calc is not None:
        combined_atoms.calc = calc

    return combined_atoms


def solutions_to_atoms_list(
    solutions,
    founder_atoms,
    fixed_atoms,
    fixed_indices,
    free_atoms,
    calc,
    frozen_atoms=None,
    frozen_indices=None,
    free_indices=None,
):
    """
    Build full Atoms for each sample by overwriting free_indices (from samples)
    and frozen_indices (from optimized frozen atoms of current gen).
    """

    if free_indices is None:
        if free_atoms.info is None or "indices" not in free_atoms.info:
            raise ValueError(
                "free_indices not provided and not in free_atoms.info"
            )
        free_indices = free_atoms.info["indices"]

    atoms_list = []
    for sample in solutions:
        free_subset = sample_to_atoms(sample, free_atoms, free_indices)
        combined = combine_fixed_frozen_and_free_atoms(
            founder_atoms=founder_atoms,
            fixed_atoms=fixed_atoms,
            fixed_indices=fixed_indices,
            free_atoms=free_subset,
            frozen_atoms=frozen_atoms,
            frozen_indices=frozen_indices,
            calc=calc,
        )
        atoms_list.append(combined)
    return atoms_list


def calculate_atoms_list(
    atoms_list, func, desc, multiproc, n_proc, progress_bar, kwargs
):
    calculated_atoms_list = []

    if progress_bar:
        if multiproc:
            with (
                Pool(processes=n_proc) as pool,
                tqdm(total=len(atoms_list), desc=desc) as pbar,
            ):
                for calculated_atoms in pool.imap(
                    partial(func, **kwargs), atoms_list
                ):
                    calculated_atoms_list.append(calculated_atoms)
                    pbar.update()
                    pbar.refresh()
        else:
            for atoms in tqdm(atoms_list, desc=desc):
                calculated_atoms_list.append(func(atoms, **kwargs))
    else:
        if multiproc:
            with Pool(processes=n_proc) as pool:
                for calculated_atoms in pool.imap(
                    partial(func, **kwargs), atoms_list
                ):
                    calculated_atoms_list.append(calculated_atoms)
        else:
            for atoms in atoms_list:
                calculated_atoms_list.append(func(atoms, **kwargs))

    return calculated_atoms_list


def atoms_list_to_solutions(atoms_list, free_indices, dimensions=None):
    """
    Extract flattened positions for free indices across a list of Atoms.
    """

    free_atoms_list = [atoms[free_indices].copy() for atoms in atoms_list]
    positions = np.array([a.get_positions() for a in free_atoms_list])
    if dimensions is None:
        dimensions = positions.shape[1] * 3
    solutions = positions.reshape(-1, dimensions)
    return torch.from_numpy(solutions)
